refactor: separate outcome and feedback template editor
