+++ src/app/admin/institution-settings/institution-settings.component.html
@@ -17,7 +17,7 @@
             <p>Manage global learning outcomes for the institution</p>
           </div>
         </div>
-        <f-feedback-template-editor></f-feedback-template-editor>
+        <f-learning-outcome-editor></f-learning-outcome-editor>
       </div>
     </mat-tab>
     @if (overseerEnabled) {
