+++ src/app/doubtfire-angular.module.ts
@@ -238,10 +238,11 @@ import {TestAttemptService} from './api/services/test-attempt.service';
 import {ScormExtensionCommentComponent} from './tasks/task-comments-viewer/scorm-extension-comment/scorm-extension-comment.component';
 import {ScormExtensionModalComponent} from './common/modals/scorm-extension-modal/scorm-extension-modal.component';
 import {FeedbackTemplateService} from './api/services/feedback-template.service';
-import {FeedbackTemplateEditorComponent} from './common/feedback-template/feedback-template-editor.component';
+import {FeedbackTemplateEditorComponent} from './common/feedback-template-editor/feedback-template-editor.component';
+import {LearningOutcomeEditorComponent} from './common/learning-outcome-editor/learning-outcome-editor.component';
 import {TaskFeedbackTemplatesComponent} from './tasks/task-comment-composer/task-feedback-templates/task-feedback-templates.component';
-import {NestedCsvDownloadModalComponent} from './common/feedback-template/nested-csv-download-modal/nested-csv-download-modal.component';
-import {NestedCsvDownloadModalService} from './common/feedback-template/nested-csv-download-modal/nested-csv-download-modal.service';
+import {NestedCsvDownloadModalComponent} from './common/learning-outcome-editor/nested-csv-download-modal/nested-csv-download-modal.component';
+import {NestedCsvDownloadModalService} from './common/learning-outcome-editor/nested-csv-download-modal/nested-csv-download-modal.service';
 
 // See https://stackoverflow.com/questions/55721254/how-to-change-mat-datepicker-date-format-to-dd-mm-yyyy-in-simplest-way/58189036#58189036
 const MY_DATE_FORMAT = {
@@ -363,6 +364,7 @@ const MY_DATE_FORMAT = {
     ScormExtensionCommentComponent,
     ScormExtensionModalComponent,
     FeedbackTemplateEditorComponent,
+    LearningOutcomeEditorComponent,
     TaskFeedbackTemplatesComponent,
     NestedCsvDownloadModalComponent,
   ],
