+++ src/app/units/states/edit/directives/unit-tasks-editor/task-definition-editor/task-definition-editor.component.html
@@ -34,7 +34,7 @@
         <h3 class="mb-1 text-2xl font-bold tracking-tight text-gray-900">Task Learning Outcomes</h3>
         <p class="font-normal text-gray-700">Add learning outcomes for this task</p>
         <div class="bg-white border border-gray-200 rounded-xl shadow p-6">
-          <f-feedback-template-editor [context]="taskDefinition"></f-feedback-template-editor>
+          <f-learning-outcome-editor [context]="taskDefinition"></f-learning-outcome-editor>
         </div>
       </div>
     </div>
