+++ src/app/units/states/edit/edit.tpl.html
@@ -5,7 +5,7 @@
     </tab>
   </tabset><!--/tabs-->
   <unit-details-editor unit="unit" ng-if="activeTab === tabs.unitTab"></unit-details-editor>
-  <f-feedback-template-editor [context]="unit" ng-if="activeTab === tabs.learningOutcomesTab"></f-feedback-template-editor>
+  <f-learning-outcome-editor [context]="unit" ng-if="activeTab === tabs.learningOutcomesTab"></f-learning-outcome-editor>
   <unit-staff-editor ng-if="activeTab === tabs.staffTab"></unit-staff-editor>
   <unit-tutorials-manager ng-if="activeTab === tabs.tutorialsTab" [unit]="unit"></unit-tutorials-manager>
   <unit-students-editor [unit]="unit" ng-if="activeTab === tabs.studentsTab"></unit-students-editor>
